+++ src/utils/validators/filePath/isValidFilePath.test.ts
@@ -3,31 +3,34 @@ import { isValidFilePath } from './isValidFilePath';
 
 describe('how isValidFilePath works', () => {
     it('is valid relative file path', () => {
-        expect(isValidFilePath(`./hello.txt`)).toBe(true);
-        expect(isValidFilePath(`../hello.txt`)).toBe(true);
-        expect(isValidFilePath(`../../hello.txt`)).toBe(true);
-        expect(isValidFilePath(`../../../hello.txt`)).toBe(true);
-        expect(isValidFilePath(`../../../../hello.txt`)).toBe(true);
-        expect(isValidFilePath(`../../../../../a/b/hello.txt`)).toBe(true);
-        // TODO: [🧠] What about just 'hello.txt'
+        expect(isValidFilePath(`./hello.book`)).toBe(true);
+        expect(isValidFilePath(`../hello.book`)).toBe(true);
+        expect(isValidFilePath(`../../hello.book`)).toBe(true);
+        expect(isValidFilePath(`../../../hello.book`)).toBe(true);
+        expect(isValidFilePath(`../../../../hello.book`)).toBe(true);
+        expect(isValidFilePath(`../../../../../a/b/hello.book`)).toBe(true);
+        expect(isValidFilePath(`hello.book`)).toBe(true);
+        expect(isValidFilePath(`foo/hello.book`)).toBe(true);
+        expect(isValidFilePath(`foo/hello`)).toBe(true);
+        // TODO: [🧠] What about just 'hello.book'
     });
 
     it('is valid Unix absolute file path', () => {
-        expect(isValidFilePath(`/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`/a/b/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`/a/b/c/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`/a/b/c/d/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`/a/b/c/d/e/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`/a/b/c/d/e/f/hello.txt`)).toBe(true);
+        expect(isValidFilePath(`/hello.book`)).toBe(true);
+        expect(isValidFilePath(`/a/b/hello.book`)).toBe(true);
+        expect(isValidFilePath(`/a/b/c/hello.book`)).toBe(true);
+        expect(isValidFilePath(`/a/b/c/d/hello.book`)).toBe(true);
+        expect(isValidFilePath(`/a/b/c/d/e/hello.book`)).toBe(true);
+        expect(isValidFilePath(`/a/b/c/d/e/f/hello.book`)).toBe(true);
     });
 
     it('is valid Windows absolute file path', () => {
-        expect(isValidFilePath(`C:/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`D:/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`E:/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`F:/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`G:/hello.txt`)).toBe(true);
-        expect(isValidFilePath(`H:/hello.txt`)).toBe(true);
+        expect(isValidFilePath(`C:/hello.book`)).toBe(true);
+        expect(isValidFilePath(`D:/hello.book`)).toBe(true);
+        expect(isValidFilePath(`E:/hello.book`)).toBe(true);
+        expect(isValidFilePath(`F:/hello.book`)).toBe(true);
+        expect(isValidFilePath(`G:/hello.book`)).toBe(true);
+        expect(isValidFilePath(`H:/hello.book`)).toBe(true);
     });
 
     it('is NOT valid file path', () => {
@@ -37,5 +40,6 @@ describe('how isValidFilePath works', () => {
         expect(isValidFilePath(`wtf://collboard.com/`)).toBe(false);
         expect(isValidFilePath(`blob:nothing`)).toBe(false);
         expect(isValidFilePath(`blob:httpx://localhost:9977/fooo/add`)).toBe(false);
+        expect(isValidFilePath(`https://promptbook.studio/foo/bar.book`)).toBe(false);
     });
 });